refactor/renaming and encapsulating better
